train/requirements.txt
torch>=2.1.0,<=2.3.0
torchvision>=0.16.0,<=0.18.0
Pillow>=9.0.0
tqdm>=4.64.0
opencv-python-headless>=4.8.0
matplotlib>=3.5.0